db_system/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
boto3
qdrant-client
psycopg2-binary
numpy
